refactor: restructure project architecture
