+++ src/components/news/NewsItem.jsx
