+++ src/components/user/AccountDelete.jsx
@@ -10,15 +10,11 @@ import {
   DrawerTrigger,
 } from "@/components/ui/drawer";
 import { useNavigate } from "react-router-dom";
-import { useContext } from "react";
-import IsLoginContext from "@/contexts/IsLoginContext";
 
 const AccountDelete = () => {
-  const { logout } = useContext(IsLoginContext);
   const navigate = useNavigate();
 
   const handleDelete = () => {
-    logout();
     navigate("/");
   };
   return (
